Move cached directories one level down within the node cache (#587)
This is a pre-requisite for making better use of the cache.

The current cache layout puts everything that the Node buildpack wants in the cache at: $CACHE_DIR/node including custom directories.

This creates a naming issue, since a user could cache anything. We currently keep a cache signature at: $CACHE_DIR/node/signature, but if a user declared a custom cache of a signature file, these would conflict and break things.

Adding anything else to the Node cache would make this problem worse, so let's fix it.

This adds one more level where all of the cached directories can live. Now if a user wants to cache a signature file, the cache layout would look like this:

node/
  - cache/
    - signature
  - signature
so there is no longer a conflict.

Deploying this will cause a cache miss for all users since the new code is looking in a new spot, but this is a common enough occurrence that I don't expect it to be significant for anyone. To ensure that the old cache will be discarded, I've added a version to the signature file.